Switches from build-image to gvm-libs and install dependencies.
Enhances the robustness of the CI process by using GVM-libs and
installing the required dependencies, instead of relying on a pre-built
image that must be pushed before running CI tests.
This approach eliminates the need to push dependencies beforehand and
wait for a docker image to be pushed before utilizing it.

Furthermore, we have moved the smoketest.yml step to nasl.yml since it
is also NASL-based.

The build and tests now verify GVM-libs unstable as well as stable to
ensure both downwards and upwards compatibility when on stable.

Our intention is to be notified of any breaking changes.